rename: SessionMember 위치를 auth_member에서 global로 변경
